fix(certs) (#37)
Remove required certs from the jenkins-* envs config. The user agreement quiz is causing failed tests in `jenkins-blood` when the PR is on `cdis-manifest` because it is pulling the config from prod. In the future we can remove this, if we want to add integration tests for the quiz feature